Water Temple and Shadow Temple
Water Temple, Shadow Temple, and the appropriate overworld stuff to handle entering Shadow Temple are all now modeled. With some minor items to add to the shuffle details, the entire world outside of the Gerudo region and Ganon's Castle should now be modeled. With Longshot and Hover Boots obtainable in the logic, only Silver Gauntlets remain among major progression items not in the logic.