Biggoron Sword and Ganon's Castle Boss Key
Biggoron Sword/Adult Trading Sequence are now in the shuffle! A random adult trading sequence item other than Odd Potion will be placed in the item pool, Anju as an adult will give a random item, the Biggoron Sword is in the item pool (with massive shenanigans in place to make it actually work), and Biggoron will give a random item. Currently these new spots are not yet integrated into the hint system; I'll deal with that tomorrow.

I've also added some code (credit to mzxrules) to allow the Ganon's Castle Boss Key to shuffle into the rest of Ganon's Castle outside of the tower. This gives us a "validation chest" and means that even those who do trials skip are going to have to hunt through the trials for that key.

Completion of junglechief's fairy shuffle is coming tomorrow probably too but it's just too late for me to keep working... I didn't test this as much as I should either so hopefully nothing breaks horribly.