Stone/medallion shuffle
In this update, the Spiritual Stones and Medallions now shuffle amongst themselves. Instead of starting with the Light Medallion always, the player starts with a random stone or medallion with the eight dungeons with bosses each containing a random prize from the pool. The pedestal at which the Song of Time is played in vanilla always gives hints for the locations of the stones (as a kid) or medallions (as an adult). Several mechanics are changed to accommodate this; the Kokiri Kid in closed forest moves out of the way based on Deku Tree clear not specifically Kokiri Emerald (so you have to get whichever prize Deku Tree has) but a few other checks have been fixed to actually use the medallions (Prelude of Light slot, Nocturne slot, also a fix to King Zora so he no longer cares about Zora Sapphire). This feature needs substantial testing but seems to work on the surface.

The logic has also received some clean-up here. Water Temple and the Shadow Trial in Ganon's Castle both had significant inaccuracy that generally demanded more than the player really needed; that has been fixed. One GS Token on adult Zora River now has a more accurate requirement as well.