Fix for Water Temple Small Key gen failures
The algorithm just wasn't smart enough to understand that it couldn't put Boss Key in spots blocked by a bunch of keys if Lullaby was behind Water Temple and there were only two non-Lullaby blocked locations for keys to go. A check for keysanity (which is currently hard coded to false but when keysanity is implemented will be different) or lullaby is added to all locked door blocks in Water Temple because that's effectively true (anything that requires keys in WT requires many of them) and this allows the logic to understand the situation ahead of time and not work itself into a corner.