Half of fast GC flag
The actual hints themselves need put in but the regions to hint at are now defined in the locations. The removal of the barrier and the Boss Key door should work. The skip of the collapsing tower isn't programmed yet.

I also fixed the logic reqs on two Skulls to be more accurate (super low% chance they would have made impossible seeds), cleaned up the spoiler log just a bit (and added the new flag to the log), and made it so making a seed with a spoiler log puts the items in different places than making a seed without.

Lastly I updated the program's version number to 2.0 which will be accurate shortly.