Fixes Compress Args for Mac/Linux
Compress for Mac and Linux take only one argument and it automatically appends `-comp` to the compressed ROM.  The previous argument does not respect `output_dir`, looking for the uncompressed output ROM in the base dir causing a file not found error if an output directory is specified.  This change uses `rom_path` as to respect the user's set `output_dir`.