Spirit Temple Logic Update
Some routes (the ones for if you had the requirements to get the goals both as child and as adult with only 1 key) were not checking for explosives. What could happen is, you unlock the lower door as child, and you have silver gauntlets but not explosives, the randomizer expected you to be able to reach goals in the main room because you had 1 key and silver gauntlets.

But also, besides 1 key and explosives, 2 keys is enough to ensure you really have access to the main room. Except it's possible to waste your keys as child if you have out of logic explosives. So the 2-key options need to check that bombchu logic is on, because only in that case we know for sure we can use explosives again. The GS adds a Boomerang check to make sure you can really get it if you get there as child.

The Fire Arrows routes for the Map and Sun Block chests required full adult access, but this was not necessary. Din's Fire and Fire Arrows are actually functionally equivalent since child can use sticks, or adult could use the fire source. (Shopsanity, however, is a different story for a different time.)

Spirit hurt my brain.