Dodongos Cavern Logic Fix
I moved the Boss and GS rules in together with the rest of the DC logic.

The slingshot target rule needed to check the child entry requirements to make sure the slingshot was actually usable. The adult trick-jump could be forced otherwise, say if you were in here with slingshot, hammer, and dins/bow.

I believe the logic on the end of bridge chest was not needed. It seemed like it was checking for adult entry requirements so you could use the hammer on the wall, but you could just hit the switch as child and come back as adult to use the hammer. The logic here was actually preventing this scenario from being able to occur.