Update hints for MQ, tokensanity
This updates the hint system to properly use MQ locations and to consider giving hints for tokensanity only locations. The hint ratios have been rebalanced a bit; at least one junk hint should always exist as this is written but no more than 10 (these are the extremes of RNG on hint distribution; you expect an average of about 5).

I also removed a line of debug code from Patches.py that I left in on accident on the last push. Oops!