Remove misc wonder text, adjust chest size flag
This will remove the miscellaneous Wonder Text actor from Gerudo Training Grounds vanilla. In general, I cleaned up the Wonder Text removal code and removed a ton of needless stuff I was doing from back when I was handling that in a worse way.

A chest in Ganon's Castle vanilla was moved by a small amount to make it possible to open if it was upsized by a randomizer flag. I made a subtle change to only reposition the chest if that flag is in use. The check was already present anyway so this is slightly more efficient and it keeps the dungeon slightly more "pure" for the vast majority of the time that flag is not in use. Todo: look into the need to reposition that chest in Ganon's Castle MQ.

I also removed some commented out bit flag setting and improved the comments on the flags that remain. Different code that reworks how Navi prompts work (they're always c-up options, never forced interrupts) made those bit flags needless, and those bit flags caused problems with MQ.