Don't shuffle shop text for non-shop items and GS text in some settings
The non-shop items texts are not shuffled if text shuffle is excepting hints. This is so that the prices can be known. (Issue #189)

The GS Token message is from a softlock that can occur when playing with full text shuffle + shopsanity + tokensanity. When those combinations of options are set, the GS Token message won't be shuffled (Issue #111)