Forbid GS from appearing in Shops (Issue #111)
Also reverted the change with not shuffling the GS. I misunderstood the problem and it did not fix the issue. This fix is somewhat of a hot fix since I'm not sure what's causing the softlock but it seems to only happen to GS Tokens in Shops. Unsure how to fix, probably related to BGS softlock