Logic fix for DC in OHKO
So originally I was thinking you would survive the jump down by shield-dropping the bomb flower at the last moment, then rolling out, then picking it back up, and then just barely throwing it in time to blow up the wall. It was pretty hard but at least it saved me from having to account for OHKO in DC.

Then come shopsanity, and now I've have to check for Buy Deku Shield, as well as bottle or Nayru's. And for the one that you could do as adult, Buy Hylian Shield (which would need to be made progressive) and Mirror Shield as well. So I decided using a shield was dumb and hard, and we'll just require bottle (or Nayru's if the minds on excluding that from OHKO ever change).

The adult routes which carry the flower down, there's an alternative where you can backwalk with a bombflower out through the one-sided collision of the bombable wall and blow it up that way. I decided that, though that wasn't hard, it was too glitch-like to required in this case.

This includes TR's change to ban tokens from shops, but does not include my logic_tricks additions, which would have to be merged separately.