Fix for Boss Key placement in key for key spots
If the logic doesn't make sense reading it, remember De Morgan's laws. In this case not (A or B) = (not A and not B). I used that to make the code shorter. The point here is that the key for key previously had two triggers: the spot contains a small key or keys not yet placed. The second condition was not a good assumption if the Boss Key landed there and a critical Small Key was blocked by the Boss Key. This adjusts the keys not placed check to fail if Boss Key is already in such a location.