Massive cleanup to hint system
A lot of the new hints were, subjectively I suppose, quite poor. This replaces a lot of them with more clever hints. A lot of new "junk hints" were added, most of which contain information that may actually be useful for new players about a few little strategic points of play. It also fixes a ton of little issues/inconsistencies in the hint system that had accumulated over the revisions.