Rewrote 'one major item per dungeon' algorithm
Issue #222, #179

Instead of trying random configurations like is done for songs and shops, choose dungeons in a random order and place one item there. The main thing is that since there are more than 13 mostly useless major items, it should always be guaranteed to work in one pass. This might have some implications on bias that I don't fully understand yet, but this entire option has heavy impacts on bias anyways.

For the issue with the spirit colossus hands, I moved the check from being in the "major items in the dungeon" to instead add a lock to every location relevant to prevent further major item placements. While I need to add it to more locations, it makes everything handling the spirit hands cleaner. Also it avoids ugliness involved with having 2 items in spirit, since it doesn't check the count, it just places the number required and assumes correctness.