Spirit temple's unique dungeon items are not placed together
This code is a bit slower, and I'm not sure this ordering actually has an impact on the bias because of subtle ways that the algorithm works. Also this function incurs so much bias already that I'm not sure that trying to reduce tiny amounts of proposed bias is meaningful in any way. However I was told to do this under duress, so here it is.