GTG fix and another capitalization
Vanilla GTG had an issue wherein the text box in the sunken silver rupees room would display as you dove because it wasn't moved low enough. This moves the text boxes to just the necessary depth to avoid this problem. I did confirm this causes no instability in Shadow Temple. I also noticed a label that wasn't capitalized with the rest of the style updates that was in Gui.py instead of in Settings.py. I took care of that here.