Change "select all trick options" to search for all trick options
This is makes it so we don't have to enumerate every option. There some question on whether this should affect the lens option, but i'm not sure what 2 of 3 states it would toggle between. Logically it feels like it should be most/least restrictive, but the middle option is what people generally use.