Add "fast start" option
This patch adds the "fast start" option, which does the following:

1. Link will start the game with a Deku shield equipped.
Justification: The Deku shield is required in closed-forest games,
and at a cost of 40 rupees, requires some grinding to obtain. Even
in open-forest games, the shield is required to progress in the
Deku Tree and therefore needs to be purchased.

2. Link will start the game with 10 Deku sticks and 20 Deku nuts.
Justification: These items are immediately obtainable in the
Kokiri shop and outside the Deku Tree.

3. Link will start the game with 99 rupees.
Justification: Rupees are easily farmed in the opening area of
the game.

4. Link will start the game with Farore's Wind.
Justification: Farore's Wind does not affect progression but can
be a time-saver, especially with the Randomizer. Link does not
start with a magic meter, which must still be obtained to use
Farore's Wind. A junk item is added to the item pool as a
replacement.

5. Link will start the game with the Prelude of Light and Serenade
of Water learned.
Justification: These two songs do not grant access to any new
areas of the game after Link leaves the Kokiri Forest, but they
can greatly speed movement by allowing Link to warp to the Temple
of Time and Lake Hylia. If playing with shuffled ocarinas and a
closed forest, Link can now leave the forest earlier if he gains
access to an ocarina; however, this was already possible if a bomb
bag, bombchu, or silver scale was generated inside the forest.
Both songs are replaced with junk items. If playing without song-
item shuffling, two non-progression, non-priority items will be
added to the song list.