Multi-World Spoiler Fixes
Gossip Stone hints for all players are now included in the spoiler log. As a result, the player number is no longer included in the filename for multi-world spoilers.

Also added [W#] next to locations in multi-world spoilers to more easily Ctrl+F the correct location/world.

Filename logic in general cleaned up a bit.