Finish integrating new subclasses into setting list
I don't feel that converting color/sfx comboboxes is worth the hassle
right now so they remain Setting_Info definitions. Which is totally
fine. Setting_Info should be exactly the same as it was.

In general the list should be getting easier to read and understand.
That's the hope anyway.