Simplify the song placement by allowing not placing all items
Unplaced items can then be handled as appropriate. They can be either passed to the next list to be placed again with fewer restrictions, or throw an error. The items will need to be placed eventually, and if the game is unbeatable then it will throw an error regardless.

The item rules also needed to be enhanced. They weren't accurate, so attempting to place items this way could lead to bugs. The rules no ensure items are placed in the correct type.

TODO: Deku Scrub item rules are not set still, since their rule is not yet known at this point and only until after the itempool is generated. Shouldn't matter since Shop items are placed separately anyways, but for completeness, it would be nice if these rules were accurate for sanity purposes.