Prevent SoA/MoT from being locked behind a disabled location Issue #336
The issue is that the soa is not needed to place to reach any locations during generations, however it is needed to talk to hint stones. In order to ensure the stone is always reachable, SoA/MoT is required to beat the game. This shouldn't really affect generation at all, but it does add SoA/MoT to the spoiler when it's not actually needed.