Remove key for key logic. Remove unused locked door in water temple
Key locked keys are pretty strange in all locations reachable, but this has been is mostly as a hold over from lttp rando and because water temple has so many key doors. By removing the unused key door in water temple, this alleviates this issue. So key locked keys are now no longer possible. It can still happen in beatable only as was before.