Fix ability to place hints if a stone is not reachable
Before we were checking is all stones were reachable, so not we'll check for each stone individually. The required hints will always be placed in a reachable stone, but if there is no reachable stone remaining then it won't be placed, so the hint might not be guaranteed in this case. This case is so rare that I'm unsure if I should even try to handle it, if it's even possible to happen.

For non-required hints, they can be placed in a unreachable stone.