Rework of barren hints to have better coverage
It ignores major items that could never be progression and it uses woth data to infer additional items that are not required in the seed. In addition a weight is added to the barren areas based on their size so that larger areas are more likely to be selected since they are less likely to be candidates. This overall adds a lot more locations available to barren (like for 3-4 to 5-10) and prevents common areas from showing up frequently.