Remove `logic_ignored_items`
So the way this was implemented does work. It assumes you have the item even if it's impossible to obtain, which means that it can place required items in unreachable locations. This options is also fairly weak in general, since if the plando maker has a deep enough understanding of logic to be able to set up weird options like this, then they should be placing item in the `no logic` mode anyways, which was the entire purpose of that option.