Remove hookshot assumptions from Forest
The first should include explosvies for completness. Change over to
has_projectile to simplify. Currently no logic impact as Link always
has a hook just being here and r0bd0g mentioned that logic is relying
on this assumption.

Qualify the "hookshot" chest and its skull with hookshot along with
backdoor access from the falling room. The skull can be killed with
a variety of weapons from above.

All needed for entrance rando.