Add new state helper functions
Add can_use_projectile as a simpler version of has_projectile that checks
for any projectile (like an oring of can_use()). Avoid changing the existing
has_projectile as that is complicated by non-local spirit temple logic.

Add a shorthand for having a usable shield (for deku scrubs) and for being
able to do fire from a torch.