Tweak DC cross age logic
Once smashed, the entryway to DC is ope cross age. Capture this. One
the other end, bombing the stairs in DC is not persistant so this
should be pulled outside the as_either.

I don't think this causes any bugs as the facts that can lead to a
situation where these expressions come into play end up defeating the
whole dungeon (i.e. this whole thing is in logic as adult anyway when
adult has a hammer and bow+trick).

But adjust it for clarity and proof against further trick definitions.