Allow Multiple 'Bottle with Letter'
If King Zora has been moved or if you have a bottled letter in any of the 4 bottle slots, you are given an empty bottle instead, with letter bottle icon/model and changed text explaining that the bottle is emptied.

Plentiful item pool updated to add a duplicate 'Bottle with Letter'.