Rewrite the collect_locations algorithm.
Playthrough is a new class that iteratively expands spheres,
saving each sphere as a region layer. It is faster than the old
algorithm for outward iteration (as in can_beat_game and
update_required_items). It is not yet faster for inward iteration
(as in create_playthrough), which is by far the majority of the
generation time.