Change Random MQ Distribution to favor event split
Both of these distributions will have independent random variables with an expected value of 6 MQ dungeons. However since for MQ, since we don't really care about the number of MQ dungeons, it would generally be preferred to favor a more split distribution with this binomial instead of uniform. The main difference between this and choosing 6 dungeons is that you cannot determine if the last dungeon is MQ or not after checking 11 of the dungeons.