Remove shooting gallery actor when entering the gallery as the wrong age
There are 2 different shooting galleries in logic, one for each age, but in game both are actually the same scene and room. 
In order to effectively have 2 shooting galleries, each playable as one age, we remove the actor from the room when entering as the wrong age.
This is only relevant in Entrance Shuffle situations since galleries are each only accessible as one age with vanilla entrances.