Add partial time of day logic for interiors
First implementation of time of day logic in order to fix an issue with time of day when shuffling interior entrances.

This doesn't update the entire logic to account for time of day and mostly just ensures internal regions shuffle works properly even with time of day requirements.

The implementation, along with overworld locations, will have to be completed later when dealing with Overworld ER, but for now all issues currently relevant should be handled with this.