Make interior entrance shuffle incompatible with Tower Escape Sequence
Until a better solution is found, or once the trade quest is reworked, the Tower Escape Sequence can't be used along with interior entrance shuffle because of a timers issue.
So for now, Skip Tower Escape Sequence is forced to be always enabled when interior entrances are shuffled, to avoid softlocks during the collapse.