Improve generation speed by skipping some useless playthrough checks
When reducing playthrough spheres, there is no point testing if the game is still beatable after uncollecting an item that can't ever be required, i.e. if a duplicate of that item is already obtained and it's not progressive. So we can just skip the test in this case, which improves performance by a decent amount.