Add a Magic Beans Shuffle option
Adds a Magic Bean Pack item which can be shuffled in the pool and gives you 10 beans when you obtain it. 
Also changes the bean salesman to allow him to sell a one time item for a set price if needed.

Both these changes allow a new option to be added that, when enabled, adds a 10 bean pack into the pool and turns the bean salesman into a one time seller.