Excluded are junk instead of 5 rupee. Song locations sets song_as_items
This means that if you exclude a song, it is junk like everything else, which is much more consistent for excluded in general. The only caveat then is that now there might not be a place to put songs. You must either turn on "start with warp songs" to reduce the number of songs in the pool, or turn on shuffle songs with items.