Nerf Bombchu Salesman. (#688)
Salesman now provides only Bombchus item, which with chus in logic
allows purchasing chus from shops and playing Bombchu Bowling, but
does not itself count as a renewable source. With chus not in logic,
Bombchus item has no effect.

Given that has_bombchus is thus impossible without a Bomb Bag
with chus not in logic, short-circuit the explosives check.