Improve the entrance shuffling retry to reduce generation fails
The local retry for entrance placement was limited to a group of entrances inside a pool, instead of the whole entrance pool. 
For example, this meant that the first entrance group could be placed in a way that prevented the second group from succeeding, and the retry could not handle that situation so it just failed repeatedly.

To prevent this, I moved the retry up in the algorithm to restore the whole entrance pool before attempting another placement.