Fix child lake control patching being always applied in ER
This never affected logic but the lake state was always shared between child and adult in ER, even with the "child may control lake hylia" option disabled. 
I think this was left from back when the child option wasn't a thing, and it was bound to dungeon ER instead.

It will now only be applied when the corresponding option is enabled.