Fix spirit temp flags ER bug
In vanilla, when you zone out of spirit onto the hands of the colossus
then jump down and re-enter spirit from the front door, the temporary
switch flags are preserved. This is bad for ER, as it means flicking
switches in spirit, will cause switches to also flick in the dungeon
on-spirit if you jump off the hands.

Fix it by always clearing the temp flags zoning into spirit from the
front door when dungeon ER is enabled. Zoning out to a hand then back
in will still preserve the flags as per vanilla, only the front door
is changed.

This works by clearing the temp preservation flag on the sprit entry
polygon data.