Fix wolfos child weapon checks in ER
We generally agreed that using explosives to kill wolfos is too tough to be in default glitchless logic, so I updated the logic to not include them.
This wasn't relevant before overworld ER because having explosives meant you could always get sticks from Goron City, so it never came up as an issue.

Also fixed a minor bug with the wolfos grotto chest where can_child_attack was used but that includes the Boomerang which doesn't actually damage them.