Rework DMC logic to fully support overworld ER
In Overworld ER you may not be able to reset your timer at the Darunia's Chamber and Mountain Trail crater entrances so the logic has to be updated to account for that fact and properly support 3 hearts accessibility without a tunic.

This rework splits each crater region (Lower, Upper and Central) into 2 regions: one "Limited" which is accessible from other regions in the crater, and the other region corresponding to having access to a nearby entrance (to reset your timer) or being able to use a Goron Tunic. The "Limited" region is essentially a subset of the original region which contains everything that should be accessible from other areas of the crater.