Randomize the Adult potion shop
In all-interiors and above, randomize the adult potion shop.

To avoid ambiguity in the hints distribution as well as minimizing
overworld structure impact, the randomization is restricted to placing
both entrances in the same hint region.

The carpenter will still block child zoning out, even if child comes
in the backdoor blind, but blind one-way locks are a part of high level
ER already (like zoning into a dive tunnel). So keep the vanilla
behaviour. With the placement of the front door in the same hint
region, this has very little gameplay impact anyway.