Gui options tweaks
The main focus of this is to get the detailed logic tab content
consistent, although there are a few other tweaks in here too.

Starting with detailed logic, im going for the following definition
of what belongs on the tab:

1: If it has any impact to no-logic gameplay it should not not be on
   this tab.
2: If it has no impact whatsoever to no-logic gameplay it should be
   on this tab.

The is because this tab greys out in entirety depending on whether
no-logic is selected or now.

This leads to the following changes:

The trade sequence clams should be booted off the tab as they do impact
no-logic gameplay. It makes sense to move to 'other' considing it is
a minor setting (affecting one check). The 'Misc' column is getting
long so buddy these two up with the Ice Traps and Item Pool to make a
consolidated column ('Item Pool') for what is in the item pool.

Night skull hunting should go to the detailed logic tab. On its face
its a "timesaver" but it does however have a massive impact to fill
and logic, with Suns appearing on WOTH due to this setting and fill
being restricted accordingly. No other timesaver does this. So pull
in to detailed logic.

ALR should be a detailed logic setting. This is the one I am a little
on the fence about, however support folks do talk about how ALR is
an un-intuitive setting that does trap up new players, so from that
perspective I can see pulling it into logic making sense. It does
also fit the definition above on detailed logic setting. The thing
I am uncertain about is it losing its randomization on random settings.

Layout wise, rather than having Lens have its own one-option setting,
just combine the 3 misc options for the detailed logic tab (ALR, Lens
and night skulls) on the headline section. ALR is the most major
setting so list it first.

I need help with how to space 3 columns evenly on the 'Other' tab.

Two other changes:

Starting age is promoted to main rules under world. When combined with
closed door, this is a major impacter to logic and a very different
early game.

The logic ruleset is promoted to main-rules headline section. This
makes the randomize main rules toggle consistent, as now all the
options under the sub-tabs are consistenly greyed out in entirety. This
is such a major setting, it is an appropriate headline (and also
dictates the grey out of the detailed logic tab). Settings
Randomization is just below it.