Cosmetics page simplification and randomized cosmetics
Simplify the cosmetics page to be a similar layout to the other pages.
This had too many micro-sections with only a few headings that have
little score for future expansion.

Wragle all sword, tunic and guantlet settings under 'equipment'.

Move the 'general' pieces to the headline as these are more about
functionality and gameplay rather than cosmetic. For these reason,
they should be exempt from randomization so put them alongside the
new randomization setting in the headline.

Music and SFX has a headline with a randomize all toggle.