Simplify shared region of MQ spirit
LS + Silvers is just a longhand of is_adult here. A step further,
has_bow + is_adult is just can_use(Bow).

This gets the logic in the new notation and prepares for adding the
colossus hand exit which doesn't mix well with this style (would
create inconsitencies in the schema use).