Fix Ocarina Songs as Fanfares on successive generations
Ocarina Songs were added to the static list of fanfares causing duplicates to appear in the list and fail the assertion that there will only ever be one copy of a fanfare at any time in successive generations using the Generation Count feature.

A new list is now created and is extended to add both fanfares and ocarina songs to the new list and then operate on that new list instead of the static ones.