Fix Wallet Logic for High Shop Prices
There is now a case to handle the logic if somebody uses plandomizer to give a shopsanity item a price over 500 rupees. Wallet logic is now also applied to Deku Scrubs, though this is just a matter of future-proofing as there is not currently an option or a way to plandomize scrub prices high enough to require a wallet.