Updated default settings
The default settings have been tweaked to better match a preset for beginners who may not want to adjust the settings. The main objective is the create a set of settings that try to open the game up as slowly as possible so that the player is much less likely to get stuck trying to find an unchecked location in a massive search space. This is done by making the game more closed. The sword is shuffled to pair with closed forest (since that option is more applicable when forest is open). The egg is shuffled so that closed DoT is not trivialized. Shuffling the songs as item or the ocarinas could have been done as well, but they are less preferred by the community. Also change hints to be always accessible and to be clearer (since new players would have trouble deciphering the normal hint messages). Other options generally tried to stick with vanilla behaviors so that they behave in the most expected way for someone who either did not read or does not understand the behavior of various settings.

To pair with this change, the previous "Beginner Preset" is now called "Easy Mode". This is because this preset has a much stronger emphasis on making a shorter seed. However since it has a lot of deviations from vanilla, it requires more knowledge of the behavior of the randomizer going in.