Changed so can't place hints in an area that has a barren hint already
This means that barren areas can still get hints if there's no barren hint. However this is only fully guaranteed if all of the barren hints are given before the other hints, which is only true for tournament dist. For other hint dist, this will reduce the odds of getting 'sometimes' hints in barren hinted areas, but is not guaranteed.