Consider Light Arrows location hinted if Ganondorf hint can be reached.
This makes it so you won't see a Way of the Hero (or a sometimes hint) pointing directly to the Light Arrows location specifically if Ganondorf is reachable logically without Light Arrows. Any items that are only needed for Light Arrows are still fair game to be hinted, of course. This just makes hint logic consider that Ganondorf as a hint.