Minor fixes to 'Ice Trap Appearance'
'Ice Trap Appearance' being set to 'Anything' should draw from the entire item pool except for other Ice Traps. Adds the "model" of chest Ice Traps to the spoiler when Chest Size Matches Contents is on. Also, "junk_items" should be a copy of "remove_junk_items" and not directly modify it.