Hints: prefer unreachable duplicates (fixes #950).
This mostly affects hint distributions that duplicate hints,
particularly tournament hint distribution.

This removes some bias in selecting where a given hint's duplicates may go;
previously a hint for a progress item would skip stones that required
that item to reach, until a reachable stone was found, and then select
the next stone. The effect of this was to make it more likely that a
hint locked behind a hinted item was not a duplicate of that other hint.

This change makes it more evenly likely that the duplicate of an item hint
(for an item not yet possessed) is at any one other (unread) stone.

My hypothesis is that this will very slightly increase early hint variety.