Revert "Update chests.c to use looks_like_item for Chest Size Matches Contents"
This reverts commit 7f890241740f2d9c4fdb63ace7b73422412803fd.

This commit caused the dungeon info menu display to break (#961), which
may cause a crash on console. Reverting this change fixes that. It is
still possible to control which items ice traps look like, but all ice
traps will still be considered "major" items for purposes of Chest Sizes
Match Content.