Plando: Quick fix to bottle count being inaccurate for open_fountain
Fixes only accounting for 3 bottles instead of 4 bottles when open_fountain is enabled.

Added docstring giving a more detailed view of each of the params passed to the fill method. I feel like this is currently the easiest way to document the item_pools for when I manipulate them by list index. 

Other aspects are described to the best of my understanding. Review and updating of this docstring is encouraged and recommended.