Rework the ToT altar text to provide more information
Adds text on the adult altar about rainbow bridge requirements and ganon's boss key location in the current seed. This is meant to provide information about these critical settings in situations where they may be randomized or unknown.
This information is always available to the player, even when the dungeon reward hints are absent under the "Maps and Compasses Give Information" setting.

Also rewords the child altar text to not mention the 3 spiritual stones as being required to open the Door of Time, since that's usually not true in rando.