Lower progression limits for item requirements.
Mainly, determine the largest Skulltula token check per-world
and set that as the limit.

This results in a ~7% performance improvement for a 5-world seed,
and negligible difference for a 1-world shuffle-tokens seed.

(We could obviously do even better by checking what's *placed* at those
locations, but this will do for now.)