Plando: Respect item_pool definitions, fix ice traps as junk
Ice Traps should not be considered junk if junk_ice_traps is off. The #Junk item group always includes them, so handle it specifically when used.

If using a plando, the item_pool definitions should be respected when automatically adding them to the pool. If there is no more junk that can be added to the pool without exceeding these defined values an exception should be raised.

The normal settings also use this code-path, so if a distribution file is not enabled the item_pool value being exceeded should not raise an exception.