Abbreviations and several fixes
idk how to rename the deku thater checks without breaking shit. they should be called forest stage for consistency and be prefaced by LW.

I did "Wasteland" and "Colossus" instead of "HW" and "DC" respectively because "HW" looks weird and "DC" is too confusing, despite not having abbreviated the dungeons.

I haven't abbreviated the gossip stone locations yet. idk if that looks better.

To do: remove cardinal directions in dungeon check names. will start on that in an hour or so