Do alter_pool before adding pending_junk_items
My intention for doing this is to make it possible to set the number of
Triforce Pieces in the world to a value below the forced value given by
multiplying based off of the item_pool_value. However, this change
should be fairly universal.

Prior to this change an exception would be thrown as the
pending_junk_items check would run out of junk to remove before the
plando'd item pool value was forced, forcing you to choose settings that
allowed all of the Triforce Pieces to be placed before reducing the
number that get placed in the world.

Now alter_pool will set the proper value, some additional checks are
done to ensure the value doesn't get exceeded by the pending_junk_items,
and then the pending_junk_items are handled.