Reset the distribution on retry (fixes #869)
Because the distribution is used for spoiler output, when we generate
the item pool, we save it in the distribution. This is a problem when
retrying, because we generate a new item pool, and this could have more
shop items in it than shop locations.

So, to fix that, we split the Distribution update method (only ever
called on init) into once-only init and once-per-retry reset. This was
already clearing the WorldDistributions to start with.

And then remember to forward disabled locations back into the distribution.