Ensure settings lists have enough bitwidth
Fixes #1052

When generating settings strings, list settings such as the Starting Inventory Items use [0...0] and [1...1] as special terminating values. However, this was not accounted for in the bitwidth of the setting, resulting in errors for settings with exactly 2^n or 2^n-1 choices, for any n. This fixes the issue by explicitly considering the special terminators when calculating the bitwidth, adding an extra bit only when needed.