Triforce Hunt: Round number of Triforce Pieces to next whole number.
Default behavior for Python when rounding a .5 is to round it to the next even number (up or down) instead of always rounding up to the next whole number.

Example, previously a goal of 3 pieces on Balanced item pool (4.5 pieces) would result in 4 pieces in the pool. Now it would result in 5.