Add option to shuffle songs in dungeon reward locations
Songs will only appear at the boss heart contains in major dungeons, or at final reward in minor dungeons. Because this is still short 1 location, the ZL location is added as the final 12th. The idea is for it to pair with the potential option in the future to start with the ZL location.